move entities from Models into Entities directory.
