+++ Models/Entities/Insurance.cs
@@ -1,3 +1,3 @@
-namespace BikeServiceAPI.Models;
+namespace BikeServiceAPI.Models.Entities;
 
 public record Insurance(Bike bike, int Price, DateTime Start, DateTime End);
