+++ Models/Entities/LanguageElement.cs
@@ -1,5 +1,5 @@
 using BikeServiceAPI.Enums;
 
-namespace BikeServiceAPI.Models;
+namespace BikeServiceAPI.Models.Entities;
 
 public record LanguageElement(Language Language, Page Page, string Location, string TextContent);
