+++ Services/BikeService.cs
@@ -1,4 +1,5 @@
 ﻿using BikeServiceAPI.Models;
+using BikeServiceAPI.Models.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace BikeServiceAPI.Services;
